requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cssselect2==0.8.0
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.5
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-extensions==4.1
django-import-export==4.3.9
django-rosetta==0.10.2
django-unfold==0.64.2
django-webpack-loader==3.2.1
fonttools==4.59.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==11.3.0
polib==1.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.1.0
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
